requirements.txt
absl-py==2.1.0
altgraph==0.17.4
appdirs==1.4.4
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cssselect==1.2.0
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
Django==5.0.6
django-cors-headers==4.3.1
django-crispy-forms==2.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
executing==2.0.1
fake-useragent==1.5.1
filelock==3.14.0
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.66.1
h5py==3.11.0
idna==3.7
importlib_metadata==7.1.0
install==1.3.5
intel-openmp==2021.4.0
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.5.0
kiwisolver==1.4.7
libclang==18.1.1
lxml==5.2.2
lxml_html_clean==0.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.4.1
mpmath==1.3.0
mysql==0.0.3
mysql-connector-python==8.4.0
mysqlclient==2.2.4
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
opencv-python==4.10.0.82
openpyxl==3.1.3
opt-einsum==3.3.0
optree==0.12.1
packaging==24.0
pandas==2.2.2
parse==1.20.1
parso==0.8.4
pdf2image==1.17.0
pefile==2023.2.7
pillow==10.3.0
platformdirs==4.2.2
poppler-utils==0.1.0
prompt_toolkit==3.0.47
protobuf==4.25.5
psutil==6.0.0
pure-eval==0.2.2
pyee==11.1.0
Pygments==2.18.0
pyinstaller==6.7.0
pyinstaller-hooks-contrib==2024.6
PyJWT==2.8.0
PyMySQL==1.1.1
pyparsing==3.1.4
pyppeteer==2.0.0
pyquery==2.0.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-magic==0.4.27
python-magic-bin==0.4.14
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
pyzmq==26.0.3
regex==2024.5.15
requests==2.32.3
requests-html==0.10.0
rich==13.8.1
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
setuptools==70.0.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
stripe==9.10.0
sympy==1.12.1
tbb==2021.12.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
threadpoolctl==3.5.0
torch==2.3.1
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.1
tzdata==2024.1
urllib3==1.26.18
w3lib==2.2.0
wcwidth==0.2.13
websockets==10.4
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
zipp==3.19.2
django-redis==5.4.0
